+++ av/container/output.pyx
@@ -1,5 +1,6 @@
 from fractions import Fraction
 import logging
+import os
 
 from av.codec.codec cimport Codec
 from av.container.streams cimport StreamContainer
@@ -162,13 +163,17 @@ cdef class OutputContainer(Container):
             stream._finalize_for_output()
 
         # Open the output file, if needed.
-        cdef char *name = "" if self.file is not None else self.name
+        cdef bytes name_obj = os.fsencode(self.name if self.file is None else "")
+        cdef char *name = name_obj
         if self.ptr.pb == NULL and not self.ptr.oformat.flags & lib.AVFMT_NOFILE:
             err_check(lib.avio_open(&self.ptr.pb, name, lib.AVIO_FLAG_WRITE))
 
         # Copy the metadata dict.
-        dict_to_avdict(&self.ptr.metadata, self.metadata, clear=True,
-                       encoding=self.metadata_encoding, errors=self.metadata_errors)
+        dict_to_avdict(
+            &self.ptr.metadata, self.metadata,
+            encoding=self.metadata_encoding,
+            errors=self.metadata_errors
+        )
 
         cdef _Dictionary all_options = Dictionary(self.options, self.container_options)
         cdef _Dictionary options = all_options.copy()