+++ av/video/codeccontext.pyx
@@ -1,13 +1,12 @@
 from libc.stdint cimport int64_t
-
 cimport libav as lib
 
 from av.codec.context cimport CodecContext
+from av.error cimport err_check
 from av.frame cimport Frame
 from av.packet cimport Packet
 from av.utils cimport avrational_to_fraction, to_avrational
-from av.error cimport err_check
-from av.video.format cimport get_video_format, VideoFormat
+from av.video.format cimport VideoFormat, get_video_format
 from av.video.frame cimport VideoFrame, alloc_video_frame
 from av.video.reformatter cimport VideoReformatter
 
@@ -34,22 +33,19 @@ cdef class VideoCodecContext(CodecContext):
 
         cdef VideoFrame vframe = input
 
-        if not self.reformatter:
-            self.reformatter = VideoReformatter()
-
         # Reformat if it doesn't match.
         if (
             vframe.format.pix_fmt != self._format.pix_fmt or
             vframe.width != self.ptr.width or
             vframe.height != self.ptr.height
         ):
-            vframe.reformatter = self.reformatter
-            vframe = vframe._reformat(
+            if not self.reformatter:
+                self.reformatter = VideoReformatter()
+            vframe = self.reformatter.reformat(
+                vframe,
                 self.ptr.width,
                 self.ptr.height,
-                self._format.pix_fmt,
-                lib.SWS_CS_DEFAULT,
-                lib.SWS_CS_DEFAULT
+                self._format,
             )
 
         # There is no pts, so create one.