+++ av/video/frame.pyx
@@ -1,10 +1,11 @@
 from libc.stdint cimport uint8_t
 
-from av.deprecation import renamed_attr
-from av.enums cimport define_enum
+from av.enum cimport define_enum
 from av.error cimport err_check
-from av.video.format cimport get_video_format, VideoFormat
+from av.video.format cimport VideoFormat, get_video_format
 from av.video.plane cimport VideoPlane
+
+from av.deprecation import renamed_attr
 
 
 cdef object _cinit_bypass_sentinel
@@ -19,15 +20,15 @@ cdef VideoFrame alloc_video_frame():
     return VideoFrame.__new__(VideoFrame, _cinit_bypass_sentinel)
 
 
-PictureType = define_enum('PictureType', (
-    ('NONE', lib.AV_PICTURE_TYPE_NONE),
-    ('I', lib.AV_PICTURE_TYPE_I),
-    ('P', lib.AV_PICTURE_TYPE_P),
-    ('B', lib.AV_PICTURE_TYPE_B),
-    ('S', lib.AV_PICTURE_TYPE_S),
-    ('SI', lib.AV_PICTURE_TYPE_SI),
-    ('SP', lib.AV_PICTURE_TYPE_SP),
-    ('BI', lib.AV_PICTURE_TYPE_BI),
+PictureType = define_enum('PictureType', __name__, (
+    ('NONE', lib.AV_PICTURE_TYPE_NONE, "Undefined"),
+    ('I', lib.AV_PICTURE_TYPE_I, "Intra"),
+    ('P', lib.AV_PICTURE_TYPE_P, "Predicted"),
+    ('B', lib.AV_PICTURE_TYPE_B, "Bi-directional predicted"),
+    ('S', lib.AV_PICTURE_TYPE_S, "S(GMC)-VOP MPEG-4"),
+    ('SI', lib.AV_PICTURE_TYPE_SI, "Switching intra"),
+    ('SP', lib.AV_PICTURE_TYPE_SP, "Switching predicted"),
+    ('BI', lib.AV_PICTURE_TYPE_BI, "BI type"),
 ))
 
 
@@ -64,12 +65,6 @@ cdef useful_array(VideoPlane plane, unsigned int bytes_per_pixel=1):
 
 
 cdef class VideoFrame(Frame):
-
-    """A frame of video.
-
-    >>> frame = VideoFrame(1920, 1080, 'rgb24')
-
-    """
 
     def __cinit__(self, width=0, height=0, format='yuv420p'):
 
@@ -129,163 +124,23 @@ cdef class VideoFrame(Frame):
             id(self),
         )
 
-    def reformat(self, width=None, height=None, format=None, src_colorspace=None, dst_colorspace=None):
-        """reformat(width=None, height=None, format=None, src_colorspace=None, dst_colorspace=None)
-
-        Create a new :class:`VideoFrame` with the given width/height/format/colorspace.
-
-        :param int width: New width, or ``None`` for the same width.
-        :param int height: New height, or ``None`` for the same height.
-        :param str format: New format, or ``None`` for the same format; see :attr:`VideoFrame.format`.
-        :param str src_colorspace: Current colorspace.
-        :param str dst_colorspace: Desired colorspace.
-
-        Supported colorspaces are currently:
-            - ``'itu709'``
-            - ``'fcc'``
-            - ``'itu601'``
-            - ``'itu624'``
-            - ``'smpte240'``
-            - ``'default'`` or ``None``
-
-        """
-
-        cdef VideoFormat video_format = VideoFormat(format or self.format)
-
-        colorspace_flags = {
-            'itu709': lib.SWS_CS_ITU709,
-            'fcc': lib.SWS_CS_FCC,
-            'itu601': lib.SWS_CS_ITU601,
-            'itu624': lib.SWS_CS_SMPTE170M,
-            'smpte240': lib.SWS_CS_SMPTE240M,
-            'default': lib.SWS_CS_DEFAULT,
-            None: lib.SWS_CS_DEFAULT,
-        }
-        cdef int c_src_colorspace, c_dst_colorspace
-        try:
-            c_src_colorspace = colorspace_flags[src_colorspace]
-        except KeyError:
-            raise ValueError("Invalid src_colorspace %r" % src_colorspace)
-        try:
-            c_dst_colorspace = colorspace_flags[dst_colorspace]
-        except KeyError:
-            raise ValueError("Invalid dst_colorspace %r" % dst_colorspace)
-
-        return self._reformat(width or self.width, height or self.height, video_format.pix_fmt, c_src_colorspace, c_dst_colorspace)
-
-    cdef _reformat(self, int width, int height, lib.AVPixelFormat dst_format, int src_colorspace, int dst_colorspace):
-
-        if self.ptr.format < 0:
-            raise ValueError("Frame does not have format set.")
-
-        cdef lib.AVPixelFormat src_format = <lib.AVPixelFormat> self.ptr.format
-
-        # Shortcut!
-        if (
-            dst_format == src_format and
-            width == self.ptr.width and
-            height == self.ptr.height and
-            dst_colorspace == src_colorspace
-        ):
-            return self
-
-        # If we don't have a SwsContextProxy, create one.
-        if not self.reformatter:
-            self.reformatter = VideoReformatter()
-
-        # Try and reuse existing SwsContextProxy
-        # VideoStream.decode will copy its SwsContextProxy to VideoFrame
-        # So all Video frames from the same VideoStream should have the same one
-        with nogil:
-            self.reformatter.ptr = lib.sws_getCachedContext(
-                self.reformatter.ptr,
-                self.ptr.width,
-                self.ptr.height,
-                src_format,
-                width,
-                height,
-                dst_format,
-                lib.SWS_BILINEAR,
-                NULL,
-                NULL,
-                NULL
-            )
-
-        # We want to change the colorspace transforms. We do that by grabbing
-        # all of the current settings, changing a couple, and setting them all.
-        # We need a lot of state here.
-        cdef const int *inv_tbl
-        cdef const int *tbl
-        cdef int src_range, dst_range, brightness, contrast, saturation
-        cdef int ret
-        if src_colorspace != dst_colorspace:
-
-            with nogil:
-
-                # Casts for const-ness, because Cython isn't expressive enough.
-                ret = lib.sws_getColorspaceDetails(
-                    self.reformatter.ptr,
-                    <int**>&inv_tbl,
-                    &src_range,
-                    <int**>&tbl,
-                    &dst_range,
-                    &brightness,
-                    &contrast,
-                    &saturation
-                )
-
-            # I don't think this one can actually come up.
-            if ret < 0:
-                raise ValueError("Can't get colorspace of current format.")
-
-            with nogil:
-
-                # Grab the coefficients for the requested transforms.
-                # The inv_table brings us to linear, and `tbl` to the new space.
-                if src_colorspace != lib.SWS_CS_DEFAULT:
-                    inv_tbl = lib.sws_getCoefficients(src_colorspace)
-                if dst_colorspace != lib.SWS_CS_DEFAULT:
-                    tbl = lib.sws_getCoefficients(dst_colorspace)
-
-                # Apply!
-                ret = lib.sws_setColorspaceDetails(self.reformatter.ptr, inv_tbl, src_range, tbl, dst_range, brightness, contrast, saturation)
-
-            # This one can come up, but I'm not really sure in what scenarios.
-            if ret < 0:
-                raise ValueError("Can't set colorspace of current format.")
-
-        # Create a new VideoFrame.
-        cdef VideoFrame frame = alloc_video_frame()
-        frame._copy_internal_attributes(self)
-        frame._init(dst_format, width, height)
-
-        # Finally, scale the image.
-        with nogil:
-            lib.sws_scale(
-                self.reformatter.ptr,
-                # Cast for const-ness, because Cython isn't expressive enough.
-                <const uint8_t**>self.ptr.data,
-                self.ptr.linesize,
-                0,  # slice Y
-                self.ptr.height,
-                frame.ptr.data,
-                frame.ptr.linesize,
-            )
-
-        return frame
-
     @property
     def planes(self):
         """
-        A tuple of :class:`~av.video.plane.VideoPlane` objects.
-
-        :type: tuple
+        A tuple of :class:`.VideoPlane` objects.
         """
         # We need to detect which planes actually exist, but also contrain
         # ourselves to the maximum plane count (as determined only by VideoFrames
         # so far), in case the library implementation does not set the last
         # plane to NULL.
-        cdef int max_plane_count = self.format.ptr.nb_components
+        cdef int max_plane_count = 0
+        for i in range(self.format.ptr.nb_components):
+            count = self.format.ptr.comp[i].plane + 1
+            if max_plane_count < count:
+                max_plane_count = count
+        if self.format.name == 'pal8':
+            max_plane_count = 2
+
         cdef int plane_count = 0
         while plane_count < max_plane_count and self.ptr.extended_data[plane_count]:
             plane_count += 1
@@ -301,25 +156,50 @@ cdef class VideoFrame(Frame):
         def __get__(self): return self.ptr.height
 
     property key_frame:
-        """Is this frame a key frame?"""
+        """Is this frame a key frame?
+
+        Wraps :ffmpeg:`AVFrame.key_frame`.
+
+        """
         def __get__(self): return self.ptr.key_frame
 
     property interlaced_frame:
-        """Is this frame an interlaced or progressive?"""
+        """Is this frame an interlaced or progressive?
+
+        Wraps :ffmpeg:`AVFrame.interlaced_frame`.
+
+        """
         def __get__(self): return self.ptr.interlaced_frame
 
     @property
     def pict_type(self):
+        """One of :class:`.PictureType`.
+
+        Wraps :ffmpeg:`AVFrame.pict_type`.
+
+        """
         return PictureType.get(self.ptr.pict_type, create=True)
 
     @pict_type.setter
     def pict_type(self, value):
         self.ptr.pict_type = PictureType[value].value
 
+    def reformat(self, *args, **kwargs):
+        """reformat(width=None, height=None, format=None, src_colorspace=None, dst_colorspace=None, interpolation=None)
+
+        Create a new :class:`VideoFrame` with the given width/height/format/colorspace.
+
+        .. seealso:: :meth:`.VideoReformatter.reformat` for arguments.
+
+        """
+        if not self.reformatter:
+            self.reformatter = VideoReformatter()
+        return self.reformatter.reformat(self, *args, **kwargs)
+
     def to_rgb(self, **kwargs):
         """Get an RGB version of this frame.
 
-        Any ``**kwargs`` are passed to :meth:`VideoFrame.reformat`.
+        Any ``**kwargs`` are passed to :meth:`.VideoReformatter.reformat`.
 
         >>> frame = VideoFrame(1920, 1080)
         >>> frame.format.name
@@ -333,7 +213,7 @@ cdef class VideoFrame(Frame):
     def to_image(self, **kwargs):
         """Get an RGB ``PIL.Image`` of this frame.
 
-        Any ``**kwargs`` are passed to :meth:`VideoFrame.reformat`.
+        Any ``**kwargs`` are passed to :meth:`.VideoReformatter.reformat`.
 
         .. note:: PIL or Pillow must be installed.
 
@@ -360,16 +240,19 @@ cdef class VideoFrame(Frame):
     def to_ndarray(self, **kwargs):
         """Get a numpy array of this frame.
 
-        Any ``**kwargs`` are passed to :meth:`VideoFrame.reformat`.
+        Any ``**kwargs`` are passed to :meth:`.VideoReformatter.reformat`.
 
         .. note:: Numpy must be installed.
+
+        .. note:: For ``pal8``, an ``(image, palette)`` tuple will be returned,
+        with the palette being in ARGB (PyAV will swap bytes if needed).
 
         """
         cdef VideoFrame frame = self.reformat(**kwargs)
 
         import numpy as np
 
-        if frame.format.name == 'yuv420p':
+        if frame.format.name in ('yuv420p', 'yuvj420p'):
             assert frame.width % 2 == 0
             assert frame.height % 2 == 0
             return np.hstack((
@@ -385,8 +268,12 @@ cdef class VideoFrame(Frame):
             return useful_array(frame.planes[0], 3).reshape(frame.height, frame.width, -1)
         elif frame.format.name in ('argb', 'rgba', 'abgr', 'bgra'):
             return useful_array(frame.planes[0], 4).reshape(frame.height, frame.width, -1)
-        elif frame.format.name in ('gray', 'gray8'):
+        elif frame.format.name in ('gray', 'gray8', 'rgb8', 'bgr8'):
             return useful_array(frame.planes[0]).reshape(frame.height, frame.width)
+        elif frame.format.name == 'pal8':
+            image = useful_array(frame.planes[0]).reshape(frame.height, frame.width)
+            palette = np.frombuffer(frame.planes[1], 'i4').astype('>i4').reshape(-1, 1).view(np.uint8)
+            return image, palette
         else:
             raise ValueError('Conversion to numpy array with format `%s` is not yet supported' % frame.format.name)
 
@@ -395,7 +282,7 @@ cdef class VideoFrame(Frame):
     @staticmethod
     def from_image(img):
         """
-        Construct a frame from a `PIL.Image`.
+        Construct a frame from a ``PIL.Image``.
         """
         if img.mode != 'RGB':
             img = img.convert('RGB')
@@ -409,8 +296,23 @@ cdef class VideoFrame(Frame):
     def from_ndarray(array, format='rgb24'):
         """
         Construct a frame from a numpy array.
+
+        .. note:: for ``pal8``, an ``(image, palette)`` pair must be passed.
+        `palette` must have shape (256, 4) and is given in ARGB format
+        (PyAV will swap bytes if needed).
         """
-        if format == 'yuv420p':
+        if format == 'pal8':
+            array, palette = array
+            assert array.dtype == 'uint8'
+            assert array.ndim == 2
+            assert palette.dtype == 'uint8'
+            assert palette.shape == (256, 4)
+            frame = VideoFrame(array.shape[1], array.shape[0], format)
+            copy_array_to_plane(array, frame.planes[0], 1)
+            frame.planes[1].update(palette.view('>i4').astype('i4').tobytes())
+            return frame
+
+        if format in ('yuv420p', 'yuvj420p'):
             assert array.dtype == 'uint8'
             assert array.ndim == 2
             assert array.shape[0] % 3 == 0
@@ -437,7 +339,7 @@ cdef class VideoFrame(Frame):
             assert array.dtype == 'uint8'
             assert array.ndim == 3
             assert array.shape[2] == 4
-        elif format in ('gray', 'gray8'):
+        elif format in ('gray', 'gray8', 'rgb8', 'bgr8'):
             assert array.dtype == 'uint8'
             assert array.ndim == 2
         else: